Fixed event resource update
Changed events to filter with related id
Changed models to use doubles for resource values and production

+++ API/Domain/ResourceItem.cs
@@ -105,7 +105,7 @@ namespace API.Domain
         /// </summary>
         public int Level { get; set; } = 0;
         [Required]
-        public int ProductionSpeed { get; set; }
+        public double ProductionSpeed { get; set; }
 
         /// <summary>
         /// Enum used to keep track of the resource type within the API to avoid errors caused by hard-coding.
